修改精度丢失问题
# message auto-generated for no-merge-commit merge:
merge double into master

修复插入数据过程中精度缺失问题

Created-by: insightDeveloper

Match-id-237fd88a19ab806cde0144b55250ad5d018dbd7f

+++ server/src/modules/memory/parser/MemoryDef.h
@@ -13,18 +13,18 @@ namespace Memory {
 
 struct Record {
     std::string component;
-    double timesTamp;
-    double totalAllocated;
-    double totalReserved;
+    int64_t timesTamp;
+    int64_t totalAllocated;
+    int64_t totalReserved;
     std::string deviceType;
 };
 
 struct Operator {
     std::string name;
-    double size;
-    double allocationTime;
-    double releaseTime;
-    double duration;
+    int64_t size;
+    int64_t allocationTime;
+    int64_t releaseTime;
+    int64_t duration;
     std::string deviceType;
 };
